requirements.txt
appengine-python-standard>=1.0.0
google-cloud-storage
google-cloud-vision
pandas
